+++ app/controllers/groups_controller.rb
@@ -1,18 +1,46 @@
 class GroupsController < ApplicationController
-  before_action :set_group, only: %i[show edit update destroy]
+  include GroupsHelper
+  before_action :set_group, only: %i[show edit update destroy journal]
   before_action :set_department_curators, only: %i[create edit new]
 
+  def show
+    authorize @group
 
-  def show; end
+    if current_user.is_admin? || current_user.is_lecturer?
+      @students = @group.students
+      @marks = Mark.where(student_id: @students)
+    elsif current_user.is_student?
+      if @group.id == current_user.owner.group_id
+        @students = @group.students
+        @marks = Mark.where(student_id: @students)
+      else
+        @students = nil
+      end
+    else
+      @students = nil
+    end
+  end
 
   def new
     @group = Group.new
+    authorize @group
   end
 
-  def edit; end
+  def journal
+    authorize @group
+    @students = @group.students
+    @subject = Subject.where(id: params[:subject])
+    @marks = Mark.where(student_id: @students)
+    @marks = subject_marks(@marks, @subject)
+  end
+
+  def edit
+    authorize @group
+  end
 
   def create
     @group = Group.new(group_params)
+    authorize @group
 
     respond_to do |format|
       if @group.save
@@ -26,6 +54,7 @@ class GroupsController < ApplicationController
   end
 
   def update
+    authorize @group
     respond_to do |format|
       if @group.update(group_params)
         format.html { redirect_to group_url(@group), notice: 'Group was successfully updated.' }
@@ -38,34 +67,38 @@ class GroupsController < ApplicationController
   end
 
   def destroy
-
+    authorize @group
     respond_to do |format|
       if @group.destroy
         format.html { redirect_to groups_url, notice: "Group was successfully destroyed." }
         format.json { head :no_content }
       else
-        format.html { redirect_to groups_url, notice: "Something went wrong."}
+        format.html { redirect_to groups_url, notice: "Something went wrong." }
       end
     end
   end
 
   def index
     @groups = Group.includes(:department, :curator).order(:department_id).all
+    authorize Group
   end
 
   private
 
   def set_group
-    @group = Group.find(params[:id])
+    if params[:group_id]
+      @group = Group.find(params[:group_id])
+    else
+      @group = Group.find(params[:id])
+    end
   end
 
   def set_department_curators
-    @departments = Department.select(:id,:name)
+    @departments = Department.select(:id, :name)
     @curators = Lecturer.free_curators(@group).select(:id, :name)
-
   end
 
   def group_params
-    params.require(:group).permit(:department_id, :curator_id, :specialization_code, :course, :form_of_education)
+    params.require(:group).permit(:department_id, :curator_id, :specialization_code, :course, :form_of_education, :subject)
   end
 end
